// tools/jsonpath-tester/client/types.ts
export type JsonpathTesterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  jsonLabel: string
  jsonDescription: string
  jsonPlaceholder: string
  queryLabel: string
  queryDescription: string
  queryPlaceholder: string
  examplesLabel: string
  examplesPlaceholder: string
  formatJsonLabel: string
  importFromFileLabel: string
  useSampleLabel: string
  clearLabel: string
  copyQueryLabel: string
  copyResultLabel: string
  copiedLabel: string
  resultLabel: string
  resultDescription: string
  resultModeLabel: string
  matchCountLabel: string
  valuesTabLabel: string
  pathsTabLabel: string
  resultEmptyDescription: string
  noMatchesLabel: string
  downloadJsonLabel: string
  invalidJsonLabel: string
  invalidQueryLabel: string
  exampleAuthorsLabel: string
  exampleCheapBooksLabel: string
  exampleBicycleColorLabel: string
}>